+++ backend/canisters/group/impl/src/updates/delete_messages.rs
@@ -5,12 +5,11 @@ use candid::Principal;
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
 use chat_events::DeleteMessageResult;
+use constants::{MINUTE_IN_MS, OPENCHAT_BOT_USER_ID};
 use group_canister::delete_messages::{Response::*, *};
 use group_chat_core::DeleteMessagesResult;
 use types::{Achievement, CanisterId, UserId};
 use user_index_canister_c2c_client::lookup_user;
-use utils::consts::OPENCHAT_BOT_USER_ID;
-use utils::time::MINUTE_IN_MS;
 
 #[update(candid = true, msgpack = true)]
 #[trace]