+++ backend/canisters/group_index/impl/src/updates/c2c_mark_group_import_complete.rs
@@ -4,9 +4,9 @@ use crate::updates::c2c_mark_community_active::c2c_mark_community_active_impl;
 use crate::{mutate_state, RuntimeState};
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
+use constants::OPENCHAT_BOT_USER_ID;
 use group_index_canister::c2c_mark_group_import_complete::{Response::*, *};
 use types::{CommunityId, CommunityImportedInto};
-use utils::consts::OPENCHAT_BOT_USER_ID;
 
 #[update(guard = "caller_is_community_canister", msgpack = true)]
 #[trace]