+++ backend/canisters/user/impl/src/updates/approve_transfer.rs
@@ -3,10 +3,10 @@ use crate::model::pin_number::VerifyPinError;
 use crate::{mutate_state, run_regular_jobs, RuntimeState};
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
+use constants::NANOS_PER_MILLISECOND;
 use icrc_ledger_types::icrc2::approve::ApproveArgs;
 use types::TimestampNanos;
 use user_canister::approve_transfer::{Response::*, *};
-use utils::time::NANOS_PER_MILLISECOND;
 
 #[update(guard = "caller_is_owner", msgpack = true)]
 #[trace]